feat: add AccountSettingsDialog component and update auth schema
