+++ packages/shared-schemas/src/auth.schema.ts
@@ -62,6 +62,8 @@ export const userSchema = z.object({
 
 export const adminSchema = z.object({
   sub: z.string().min(1),
+  username: z.string().optional(),
+  isRoot: z.boolean().optional(),
 });
 
 /**
